requirements.txt
svgwrite
pygithub
pillow
python-dotenv